+++ src/app/(auth)/login/_components/LoginCard/index.ts
@@ -1,0 +1,1 @@
+export { LoginCard } from './LoginCard';
