--- src/app/(auth)/login/_components/RegisterCard/index.ts
@@ -1,1 +1,0 @@
-export { LoginCard as RegisterCard } from './LoginCard';
